requirements.txt
allure-pytest==2.14.2
pytest==8.3.5
selenium==4.32.0
requests==2.32.3
webdriver_manager==4.0.2
allure-python-commons==2.14.2
webdriver-manager==4.0.2